Simplify socketlabs check for status and messaging when something goes wrong.
