+++ src/olympia/lib/settings_base.py
@@ -924,7 +924,6 @@ CELERY_TASK_ROUTES = {
     'olympia.reviewers.tasks.recalculate_post_review_weight': {'queue': 'cron'},
     'olympia.users.tasks.sync_suppressed_emails_task': {'queue': 'cron'},
     'olympia.users.tasks.send_suppressed_email_confirmation': {'queue': 'devhub'},
-    'olympia.users.tasks.check_suppressed_email_confirmation': {'queue': 'devhub'},
     # Reviewers.
     'olympia.lib.crypto.tasks.sign_addons': {'queue': 'reviewers'},
     # Admin.
